move-refactor analysis.elastic.tensors to core
