+++ pymatgen/analysis/elasticity/__init__.py
@@ -2,9 +2,6 @@
 # Copyright (c) Pymatgen Development Team.
 # Distributed under the terms of the MIT License.
 
-import numpy as np
-
 from .elastic import *
-from .tensors import *
 from .stress import *
 from .strain import *
