+++ pymatgen/analysis/elasticity/stress.py
@@ -10,7 +10,7 @@ This module provides the Stress class used to create, manipulate, and
 calculate relevant properties of the stress tensor.
 """
 
-from pymatgen.analysis.elasticity.tensors import SquareTensor
+from pymatgen.core.tensors import SquareTensor
 import math
 import numpy as np
 import warnings
