+++ pymatgen/analysis/nmr.py
@@ -3,7 +3,7 @@
 # Distributed under the terms of the MIT License.
 
 from __future__ import division, unicode_literals
-from pymatgen.analysis.elasticity.tensors import SquareTensor
+from pymatgen.core.tensors import SquareTensor
 from collections import namedtuple
 
 from pymatgen.core.units import FloatWithUnit
